+++ packages/plugin-packages/mashroom-portal/type-definitions/api.ts
@@ -871,6 +871,12 @@ export interface MashroomPortalUserService {
     getAuthenticationExpiration(): Promise<number | null>;
 
     /**
+     * Get the unix ms left until authentication expiration.
+     * Returns null if the check fails and "0" if the check returns 403.
+     */
+    getTimeToAuthenticationExpiration(): Promise<number | null>;
+
+    /**
      * Extend the authentication.
      * Can be used to update the authentication when no server interaction has occurred for a while and the authentication is about to expire.
      */